edited service provider session bean and entity

+++ EasyAppointmentSystemWeb/EasyAppointmentSystemWebLibrary/src/ejb/session/stateless/ServiceProviderEntitySessionBeanRemote.java
@@ -33,6 +33,6 @@ public interface ServiceProviderEntitySessionBeanRemote
     
     public ServiceProviderEntity retrieveServiceProviderEntityByName(String name) throws ServiceProviderNotFoundException;
 
-    public void updateRating(Long newRating, Long serviceProviderId) throws ServiceProviderNotFoundException;
+    public void updateRating(double newRating, Long serviceProviderId) throws ServiceProviderNotFoundException;
 
 }
